requirements.txt
aiogram>=3.0.0
google-generativeai
pydub
python-dotenv
aiofiles
